requirements.txt
python-keyczar==0.71c
pycrypto==2.6.1
